+++ packages/core/src/api/logs/index.ts
@@ -1,5 +1,3 @@
-export { allLogLevels, defaultLogLevel, LogLevel } from './const';
-
 export { initializeLogsAPI } from './initialize';
 
 export type { LogContext, LogEvent, LogsAPI, PushLogOptions } from './types';
