+++ packages/core/src/utils/index.ts
@@ -34,8 +34,8 @@ export {
 } from './is';
 export type { IsFnHelper } from './is';
 
+export { allLogLevels, defaultLogLevel, LogLevel } from './logLevels';
+
 export { noop } from './noop';
 
 export { prefixAgentMessage } from './prefixAgentMessage';
-
-export { mockConfig, mockInternalLogger, mockStacktraceParser, MockTransport } from './tests';
