+++ packages/web/src/instrumentations/errors/registerOnerror.test.ts
@@ -1,4 +1,4 @@
-import { mockConfig, MockTransport } from '@grafana/agent-core/src/utils';
+import { mockConfig, MockTransport } from '@grafana/agent-core/src/testUtils';
 
 import { initializeGrafanaAgent } from '../../initialize';
 import { registerOnerror } from './registerOnerror';
